Data.List.RadixSort.Base

Description

Radix sort (O(k n) where k= #digits) of lists of floats (based on its IEEE754 representation) or IntN or WordN based on their representation

The lsd prefix is for Least significant digit radix sort, while msd is for the parallelized Most significant digit one.

Here we partition numbers by sign and sort both lists in parallel (you should link with -threaded)

The digit size is set to 8 bits for lists > 512 elements, else 4, in order to save space when sorting small lists

digit value queues are appended as difference lists (DList from package dlist, O(1) on append)

The instance for Int (machine word size) is not supported because it may have reserved bits for compiler use. Check The word size story. at http://www.haskell.org/ghc/docs/7.2.2/html/libraries/ghc-prim-0.2.0.0/GHC-Prim.html#g:1

A quickcheck test-suite has been added.

Used (.$) = flip ($)

See http://en.wikipedia.org/wiki/Radix_sort

Documentation

msdSortInts :: RadixRep a => [a] -> [a]

sortInts partitions between positive and negative and sort each set in parallel

O(k n) where k= #digits, for each signed bag, plus sign partition and reassembling

use this for IntN types

msdSortFloats :: RadixRep a => [a] -> [a]

sortFloats partitions between positive and negative and sort by binary repr. (exponent:mantissa) for each set in parallel, reversing the negatives list after radixSort.

O(k n) where k= #digits, for each signed bag, plus sign partition, negatives reversing and reassembling

use this for Floats and Doubles

msdSortNats :: RadixRep a => [a] -> [a]

sortNats, O(k n) where k= #digits

use this for WordN types

lsdSortInts :: RadixRep a => [a] -> [a]

lsdSortFloats :: RadixRep a => [a] -> [a]

lsdSortNats :: RadixRep a => [a] -> [a]

floatToWord :: Float -> Word32

doubleToWord :: Double -> Word64

class RadixRep t where

class to instanciate for a type to be used in radix sorts

Methods

toWordRep :: (Bits a, Integral a) => t -> a

obtain a wordN representation, as for Float: toWordRep = fromIntegral . floatToWord

sizeOf :: t -> Int

size of the type in bits

signedQual :: t -> SignedQual

Signed / Unsigned

data SignedQual

Constructors

Signed

Unsigned